feat: add edit button for preview configurations
