+++ src/navsections/HelmChartSectionBlueprint/PreviewConfigurationQuickAction.tsx
@@ -1,60 +1,44 @@
-import React from 'react';
+import {useCallback} from 'react';
 
-import {Button, Tooltip} from 'antd';
+import {Button} from 'antd';
 
-import {PlusOutlined} from '@ant-design/icons';
-
-import styled from 'styled-components';
-
-import {NewPreviewConfigurationTooltip} from '@constants/tooltips';
-
-import {SectionCustomComponentProps} from '@models/navigator';
+import {ItemCustomComponentProps} from '@models/navigator';
 
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {openPreviewConfigurationEditor} from '@redux/reducers/main';
 
-import Colors from '@styles/Colors';
-
-const SuffixContainer = styled.span`
-  display: inline-block;
-`;
-
-const ButtonContainer = styled.span`
-  display: flex;
-  align-items: center;
-  padding: 0 4px;
-  margin-right: 2px;
-  & .ant-btn-sm {
-    height: 20px;
-    width: 20px;
-  }
-`;
-
-const PreviewConfigurationNameSuffix: React.FC<SectionCustomComponentProps> = props => {
-  const {sectionInstance} = props;
-  const isSectionCollapsed = useAppSelector(state => state.navigator.collapsedSectionIds.includes(sectionInstance.id));
-
+const PreviewConfigurationQuickAction: React.FC<ItemCustomComponentProps> = props => {
+  const {itemInstance} = props;
   const dispatch = useAppDispatch();
 
-  const onClick = () => {
-    dispatch(openPreviewConfigurationEditor({helmChartId: sectionInstance.id.replace('-configurations', '')}));
-  };
+  const previewConfiguration = useAppSelector(
+    state => state.config.projectConfig?.helm?.previewConfigurationMap?.[itemInstance.id]
+  );
+
+  const helmChart = useAppSelector(state =>
+    previewConfiguration
+      ? Object.values(state.main.helmChartMap).find(h => h.filePath === previewConfiguration.helmChartFilePath)
+      : undefined
+  );
+
+  const onClickEdit = useCallback(() => {
+    if (!previewConfiguration || !helmChart) {
+      return;
+    }
+    dispatch(
+      openPreviewConfigurationEditor({helmChartId: helmChart.id, previewConfigurationId: previewConfiguration.id})
+    );
+  }, [previewConfiguration, helmChart, dispatch]);
+
+  if (!previewConfiguration || !helmChart) {
+    return null;
+  }
 
   return (
-    <SuffixContainer>
-      <ButtonContainer>
-        <Tooltip title={NewPreviewConfigurationTooltip}>
-          <Button
-            icon={<PlusOutlined />}
-            type="link"
-            onClick={onClick}
-            size="small"
-            style={{color: sectionInstance.isSelected && isSectionCollapsed ? Colors.blackPure : undefined}}
-          />
-        </Tooltip>
-      </ButtonContainer>
-    </SuffixContainer>
+    <Button onClick={() => onClickEdit()} type="link" size="small" style={{fontSize: 12}}>
+      Edit
+    </Button>
   );
 };
 
-export default PreviewConfigurationNameSuffix;
+export default PreviewConfigurationQuickAction;
